images seeder fixed, and some fixes

+++ routes/api.php
@@ -19,8 +19,6 @@ Route::group([
 ], function ($router) {
     Route::post('login', 'AuthController@login');
     Route::get('logout', 'AuthController@logout');
-    Route::post('refresh', 'AuthController@refresh');
-    Route::post('me', 'AuthController@me');
     Route::post('register','AuthController@register');
 });
 
@@ -28,12 +26,16 @@ Route::resource('galleries','GalleriesController')->except([
     'create','edit'
 ]);   
 
-Route::get('galleries/page','GalleriesController@index');
-Route::get('/authors-galleries/{id}','AuthorGalleriesController@index');
-Route::get('my-galleries','MyGalleriesController@show');
-Route::post('my-galleries/{id}','CommentsController@store');
-Route::delete('/comment/{id}','CommentsController@destroy');
-Route::delete('/gallery/{id}','GalleriesController@destroy');
+Route::get('galleries/page', 'GalleriesController@index');
+Route::middleware('auth:api')->get('authors-galleries/{id}', 'AuthorGalleriesController@index');
+Route::middleware('auth:api')->get('my-galleries', 'MyGalleriesController@show');
+Route::middleware('auth:api')->post('my-galleries/{id}', 'CommentsController@store');
+Route::middleware('auth:api')->delete('/comment/{id}', 'CommentsController@destroy');
+Route::middleware('auth:api')->delete('/gallery/{id}', 'GalleriesController@destroy');
+Route::middleware('auth:api')->get('/user', function (Request $request) {
+    return $request->user();
+});
+
